+++ gemset.nix
@@ -114,7 +114,7 @@
       path = ./.;
       type = "path";
     };
-    version = "4.4.0";
+    version = "4.4.1";
   };
   cgi = {
     groups = ["completions" "default"];